Chore: factor out ChangesBasedResolver to be shared, reusable code
