+++ shared/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod error;
 pub mod iter;
 pub mod primitives;
+pub mod resolver_utils;
